+++ app/palette_creator.hpp
@@ -25,7 +25,7 @@ namespace Texpainter
 			for(size_t k = 0; k < ret.size(); ++k)
 			{
 				ret[k] =
-				    Ui::Slider::TickMark{Ui::SliderValue{k / static_cast<double>(ret.size())}, ""};
+				    Ui::Slider::TickMark{Ui::SliderValue{static_cast<double>(k) / static_cast<double>(ret.size())}, ""};
 			}
 			return ret;
 		}
